UNR-4877: Cleanup pass
